Add the car to the Village #23
