+++ src/lib/babylon/scene/main-scene.ts
@@ -1,6 +1,7 @@
 import * as BABYLON from 'babylonjs'
 import { SceneBuilder } from '../scene-builder'
 import { CarMesh } from '../mesh/car-mesh'
+import { DwellingMesh } from '../mesh/dwelling-mesh'
 
 type AnimationKey = {
 	frame: number
@@ -15,9 +16,15 @@ export class MainScene {
 	}
 
 	private _create_car_animation(): BABYLON.Animation {
+		const car = this._scene_builder.get_by_name('car')
+		car.rotation = new BABYLON.Vector3(-Math.PI / 2, Math.PI / 2, 0)
+		car.position.y = 0.16
+		car.position.x = 3
+		car.position.z = -7
+
 		const car_animation = new BABYLON.Animation(
 			'car_animation',
-			'position.x',
+			'position.z',
 			30,
 			BABYLON.Animation.ANIMATIONTYPE_FLOAT,
 			BABYLON.Animation.ANIMATIONLOOPMODE_CYCLE
@@ -27,16 +34,21 @@ export class MainScene {
 
 		car_animation_keys.push({
 			frame: 0,
-			value: -1.5,
+			value: 7,
 		})
 
 		car_animation_keys.push({
 			frame: 150,
-			value: 1.5,
+			value: -7,
+		})
+
+		car_animation_keys.push({
+			frame: 210,
+			value: -7,
 		})
 
 		car_animation.setKeys(car_animation_keys)
-		this._scene_builder.begin_animation(['car'], car_animation, 0, 150)
+		this._scene_builder.begin_animation(['car'], car_animation, 0, 210)
 
 		return car_animation
 	}
@@ -75,9 +87,10 @@ export class MainScene {
 	public create(): this {
 		const scene_builder = this._scene_builder
 
-		scene_builder.attach_arc_rotate_camera(1.5)
-		scene_builder.attach_light({})
+		scene_builder.attach_arc_rotate_camera(15)
+		scene_builder.attach_light({ x: -1 })
 
+		new DwellingMesh(scene_builder).build()
 		new CarMesh().build()
 
 		this._create_car_animation()
